feat: Add resource to order line.
